+++ src/DebugCodeLens/index.ts
@@ -1,2 +1,3 @@
 export { DebugCodeLensProvider } from './DebugCodeLensProvider';
 export { TestState } from './TestState';
+export { DebugTestIdentifier } from './DebugCodeLens';