+++ src/TestResults/TestResultProvider.ts
@@ -1,6 +1,6 @@
 import { TestReconciler, JestTotalResults, TestFileAssertionStatus } from 'jest-editor-support';
 import { TestReconciliationState } from './TestReconciliationState';
-import { TestResult } from './TestResult';
+import { TestResult, TestResultStatusInfo } from './TestResult';
 import { parseTest } from '../TestParser';
 import * as match from './match-by-context';
 
@@ -19,6 +19,14 @@ interface SortedTestResultsMap {
   [filePath: string]: SortedTestResults;
 }
 
+const sortByStatus = (a: TestResult, b: TestResult): number => {
+  if (a.status === b.status) {
+    return 0;
+  }
+  return (
+    TestResultStatusInfo.get(a.status).precedence - TestResultStatusInfo.get(b.status).precedence
+  );
+};
 export class TestResultProvider {
   verbose: boolean;
   private reconciler: TestReconciler;
@@ -34,6 +42,35 @@ export class TestResultProvider {
   resetCache(): void {
     this.resultsByFilePath = {};
     this.sortedResultsByFilePath = {};
+  }
+
+  private groupByRange(results: TestResult[]): TestResult[] {
+    if (!results.length) {
+      return results;
+    }
+    // build a range based map
+    const byRange: Map<string, TestResult[]> = new Map();
+    results.forEach((r) => {
+      const key = `${r.start.line}-${r.start.column}-${r.end.line}-${r.end.column}`;
+      const list = byRange.get(key);
+      if (!list) {
+        byRange.set(key, [r]);
+      } else {
+        list.push(r);
+      }
+    });
+    // sort the test by status precedence
+    byRange.forEach((list) => list.sort(sortByStatus));
+
+    //merge multiResults under the primary (highest precedence)
+    const consolidated: TestResult[] = [];
+    byRange.forEach((list) => {
+      if (list.length > 1) {
+        list[0].multiResults = list.slice(1);
+      }
+      consolidated.push(list[0]);
+    });
+    return consolidated;
   }
 
   getResults(filePath: string): TestResult[] {
@@ -60,8 +97,9 @@ export class TestResultProvider {
     } catch (e) {
       console.warn(`failed to get test result for ${filePath}:`, e);
     }
-    this.resultsByFilePath[filePath] = matchResult;
-    return matchResult;
+    const testResults = this.groupByRange(matchResult);
+    this.resultsByFilePath[filePath] = testResults;
+    return testResults;
   }
 
   getSortedResults(filePath: string): SortedTestResults {