+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Please add your own contribution below inside the Master section
 Bug-fixes within the same version aren't needed
 
 ## Master
+* fully support parameterized tests in matching, diagnosis and debugging - @connectdotz
+* optimization: remove stop/start the internal jest tests process during debug - @connectdotz
 * add a new setting for "jest.jestCommandLine" that supersede "jest.pathToJest" and "jest.pathToConfig" - @connectdotz
 
 -->